修正

+++ src/main/java/com/westlake/air/pecs/domain/bean/score/IntegrateWindowMzIntensity.java
@@ -5,9 +5,9 @@ import lombok.Data;
 @Data
 public class IntegrateWindowMzIntensity {
 
-    float mz;
+    double mz;
 
-    float intensity;
+    double intensity;
 
     boolean signalFound;
 }
